refactor: rename falcon2 to just falcon
