+++ falcon/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "falcon2",
+  "name": "falcon-vis",
   "version": "0.0.0",
   "main": "build",
   "repository": "git@github.com:cmudig/falcon.git",
