+++ falcon/src/core/util.ts
