+++ package.json
@@ -1,9 +1,8 @@
 {
-  "name": "falcon2-monorepo",
+  "name": "falcon-monorepo",
   "version": "0.0.0",
   "private": true,
-  "description": "Houses the package and examples for falcon2",
-  "main": "index.js",
+  "description": "Falcon package and examples of cross-filtering",
   "repository": "git+https://github.com/cmudig/falcon.git",
   "author": "xnought <bertuccd@oregonstate.edu>",
   "license": "ISC",
@@ -15,8 +14,8 @@
     "start:flights": "yarn workspace arrow-flights dev --open",
     "jest": "NODE_OPTIONS=--experimental-vm-modules npx jest",
     "test": "yarn jest",
-    "build": "yarn workspace falcon2 build",
-    "format": "yarn workspace falcon2 format"
+    "build": "yarn workspace falcon build",
+    "format": "yarn workspace falcon format"
   },
   "devDependencies": {
     "jest": "^29.3.1",
@@ -24,10 +23,11 @@
     "typescript": "^4.9.4"
   },
   "directories": {
-    "examples": "./examples/*"
+    "examples": "./examples/*",
+    "falcon": "./falcon"
   },
   "workspaces": [
-    "./falcon2",
+    "./falcon",
     "./examples/*"
   ]
 }
